requirements_dev.txt
-r requirements.txt
selenium~=3.141
pytest~=6.2.2
pytest-django~=4.1
pytest-cov~=2.11